// src/app/modules/cheque-accounting/models/cheque-register.ts
export class ChequeRegister {
  ChqRegId: number=0;
  CompId: number=0;
  BranchId: number=0;
  StokistId: number=0;
  StockistNo:string="";
  StockistName: string="";
  StockistCode: string="";
  StockistCity: number= 0;
  ChqReceivedDate: Date = new Date();
  BankId: number = 0;
  BankName: string="";
  City: string="";
  AccountNo: number=0;
  IFSCCode: string="";
  ChqNo: string="";
  FromChqNo: string="";
  ToChqNo: string="";
  IsActive: string = '';
  Addedby: string = '';
  Updatedby: string = '';
  Action: string = '';
}